Abyss/requirements.txt
absl-py==0.11.0
addict==2.4.0
altgraph==0.17
argon2-cffi==20.1.0
astor==0.8.1
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
beautifulsoup4==4.9.3
bleach==3.2.2
bs4==0.0.1
cachetools==4.2.1
category-encoders==2.2.2
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
colorama==0.4.4
configparser==5.0.1
cycler==0.10.0
Cython==0.29.21
dataclasses==0.8
decorator==4.4.2
defusedxml==0.6.0
docker-pycreds==0.4.0
easydict==1.9
entrypoints==0.3
future==0.18.2
gast==0.3.3
gitdb==4.0.5
GitPython==3.1.14
google-auth==1.27.0
google-auth-oauthlib==0.4.2
grpcio==1.35.0
idna==2.10
imageio==2.9.0
importlib-metadata==3.4.0
ipykernel==5.4.3
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==2.11.2
joblib==0.17.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.11
jupyter-console==6.2.0
jupyter-core==4.7.0
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
lxml==4.6.1
Markdown==3.3.4
MarkupSafe==1.1.1
matplotlib==3.3.3
minepy==1.2.5
mistune==0.8.4
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.4.3
networkx==2.5
notebook==6.2.0
numpy==1.19.2
oauthlib==3.1.0
opencv-contrib-python==3.4.1.15
opencv-python==4.5.1.48
packaging==20.8
paddlepaddle-gpu==2.0.0.post110
pandas==1.1.4
pandoc==1.0.2
pandocfilters==1.4.3
parso==0.8.1
pathtools==0.1.2
patsy==0.5.1
pefile==2019.4.18
pickleshare==0.7.5
Pillow==8.0.1
playsound==1.2.2
ply==3.11
prometheus-client==0.9.0
promise==2.3
prompt-toolkit==3.0.14
protobuf==3.15.2
psutil==5.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.2
pycocotools-windows==2.0.0.2
pycparser==2.20
Pygments==2.7.4
pyinstaller==4.0
pyinstaller-hooks-contrib==2020.10
pypandoc==1.5
pyparsing==2.4.7
pyrsistent==0.17.3
PySide2==5.15.2
python-dateutil==2.8.1
pytz==2020.4
PyWavelets==1.1.1
pywin32==300
pywin32-ctypes==0.2.0
pywinpty==0.5.7
PyYAML==5.4.1
pyzmq==21.0.1
qtconsole==5.0.2
QtPy==1.9.0
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-image==0.17.2
scikit-learn==0.23.2
scipy==1.5.4
seaborn==0.11.1
selenium==3.141.0
Send2Trash==1.5.0
sentry-sdk==0.20.3
shiboken2==5.15.2
shortuuid==1.0.1
six==1.15.0
sklearn==0.0
smmap==3.0.5
soupsieve==2.0.1
statsmodels==0.12.2
subprocess32==3.5.4
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorboardX==2.1
terminado==0.9.2
testpath==0.4.4
thop==0.0.31.post2005241907
threadpoolctl==2.1.0
tifffile==2020.9.3
torch==1.7.1
torchvision==0.8.2
tornado==6.1
tqdm==4.57.0
traitlets==4.3.3
typing-extensions==3.7.4.3
urllib3==1.25.11
wandb==0.10.20
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
xgboost==1.3.3
zipp==3.4.0
